requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asn1crypto==1.5.1
attrs==25.4.0
bech32==1.2.0
bip_utils==2.10.0
bitarray==3.8.0
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.0
coincurve==20.0.0
crcmod==1.7
cytoolz==1.1.0
discord==2.3.2
discord.py==2.6.4
distro==1.9.0
ecdsa==0.19.1
ed25519-blake2b==1.4.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.121.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
grpcio-tools==1.76.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.11.1
lighter-sdk @ git+https://github.com/elliottech/lighter-python.git@c9b8de6fd6864ede0fd7fb61d85ffb3614d178c9
multidict==6.7.0
openai==2.7.1
parsimonious==0.10.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
py-sr25519-bindings==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tiktoken==0.12.0
toolz==1.1.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.0.7
uvicorn==0.38.0
v4-proto==8.2.0
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
